Laravel

+++ web_server/nginx/install.sh
@@ -2,4 +2,9 @@
 
 # install nginx
 sudo apt update -y
-sudo apt install nginx php8.1-fpm -y
+if [ -e /var/run/nginx.pid ]; then
+    echo "nginx is running";
+else
+    sudo apt install nginx php8.1-fpm -y
+    sudo systemctl enable nginx
+fi
